linking both backends to the supabase database

+++ Products/src/main/java/com/gloyoo/products/product/dto/ProductRequest.java
@@ -1,11 +1,12 @@
 package com.gloyoo.products.product.dto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record ProductRequest(
         String name,
         String description,
-        double price,
+        BigDecimal price,
         int quantity,
         String[] tags,
         Boolean active,
